requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
astunparse==1.6.3
attrs==25.4.0
bleach==6.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
distlib==0.4.0
distro==1.9.0
Django==5.1.2
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-js-asset==3.1.2
django-tailwind==4.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dm-tree==0.1.9
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.2
etils==1.13.0
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
immutabledict==4.2.2
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.0
kaggle==1.8.4
kagglesdk==0.1.15
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.7.0
mysqlclient==2.2.7
namex==0.1.0
networkx==3.5
numpy==2.2.6
openai==2.4.0
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
polars==1.34.0
polars-runtime-32==1.34.0
promise==2.3
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.1
psycopg2==2.9.10
pyarrow==23.0.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
pyparsing==3.2.5
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
scipy==1.16.2
setuptools==80.9.0
simple-parsing==0.1.8
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
termcolor==3.2.0
text-unidecode==1.3
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
twilio==9.8.7
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.218
ultralytics-thop==2.0.17
urllib3==2.5.0
virtualenv==20.34.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0



